user & sales fixes

+++ Domain/Entities/Users/User.cs
@@ -23,7 +23,7 @@ namespace Domain.Entities.Users
         public Tenant Tenant { get; set; } = null!;
         public Branch? Branch { get; set; }
         public bool IsEmailVerified { get; set; } = false;
-        public static User Create(int tenantId, string name, string email, string passwordHash, UserRole role, int? branchId)
+        public static User Create(long tenantId, string name, string email, string passwordHash, UserRole role, long? branchId)
         {
             return new User
             {
